[#160] refactor: 바뀐 모델 데이터에 따라 테스트 관련 코드 변경 및 		  바뀐 모델 이름에 따라 bnb에서 room으로 이름 수정
